move extension function out of Extensions and to the BaseFragment
