+++ app/src/main/java/com/lambdaschool/hackathon_portal/ui/fragments/BaseFragment.kt
@@ -3,8 +3,10 @@ package com.lambdaschool.hackathon_portal.ui.fragments
 import android.content.Context
 import android.os.Bundle
 import android.view.View
+import androidx.annotation.IdRes
 import androidx.fragment.app.Fragment
 import androidx.navigation.NavController
+import androidx.navigation.NavOptions
 import com.lambdaschool.hackathon_portal.ui.MainActivity
 import com.lambdaschool.hackathon_portal.viewmodel.ViewModelProviderFactory
 import javax.inject.Inject
@@ -27,6 +29,26 @@ abstract class BaseFragment: Fragment() {
     lateinit var viewModelProviderFactory: ViewModelProviderFactory
     @Inject
     lateinit var navController: NavController
+
+    /**
+     * Builds NavOptions to popUpTo and navigate to specified fragment.
+     *
+     * @param bundle: Bundle
+     * @param popUpTo_ResId : @IdRes Int
+     * @param popUpToInclusive : Boolean
+     * @param navigateTo_ResId : @IdRes Int
+     * */
+    fun navigateAndPopUpTo(bundle: Bundle,
+                           @IdRes popUpTo_ResId: Int,
+                           popUpToInclusive: Boolean,
+                           @IdRes navigateTo_ResId: Int) {
+        val navOptions = NavOptions.Builder()
+            .setPopUpTo(popUpTo_ResId, popUpToInclusive)
+            .build()
+
+        navController.navigate(navigateTo_ResId, bundle, navOptions)
+    }
+
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
