+++ app/src/main/java/com/lambdaschool/hackathon_portal/ui/fragments/create/CreateHackathonFragment.kt
@@ -15,7 +15,6 @@ import com.afollestad.materialdialogs.datetime.datePicker
 import com.lambdaschool.hackathon_portal.R
 import com.lambdaschool.hackathon_portal.model.Hackathon
 import com.lambdaschool.hackathon_portal.ui.fragments.BaseFragment
-import com.lambdaschool.hackathon_portal.util._navigateAndPopUpTo
 import com.lambdaschool.hackathon_portal.util._toastLong
 import com.lambdaschool.hackathon_portal.util._toastShort
 import kotlinx.android.synthetic.main.fragment_create_hackathon.*
@@ -85,7 +84,7 @@ class CreateHackathonFragment : BaseFragment() {
                 createHackathonViewModel.postHackathon(newHackathon).observe(this, Observer {
                     if (it != null) {
                         if (it) {
-                            navController._navigateAndPopUpTo(
+                            navigateAndPopUpTo(
                                 Bundle(), R.id.nav_user_hackathons, true, R.id.nav_user_hackathons
                             )
                             activity?._toastLong("Successfully created Hackathon")
