+++ app/src/main/java/com/lambdaschool/hackathon_portal/ui/fragments/logout/LogoutFragment.kt
@@ -46,7 +46,7 @@ class LogoutFragment : NavDrawerFragment() {
                     credentialsManager.clearCredentials()
                     wipeCurrentUser()
                     wipeLoggedInUser()
-                    navController._navigateAndPopUpTo(
+                    navigateAndPopUpTo(
                         Bundle(), R.id.nav_graph, true, R.id.nav_login
                     )
                 }
@@ -54,7 +54,7 @@ class LogoutFragment : NavDrawerFragment() {
                 override fun onFailure(error: Auth0Exception?) {
                     Log.i(TAG, "Failure ${error?.message}")
                     drawerLayout._unlockDrawer(null)
-                    navController._navigateAndPopUpTo(
+                    navigateAndPopUpTo(
                         Bundle(), R.id.nav_dashboard, true, R.id.nav_dashboard
                     )
                     activity?._toastLong("Logout Failed:\n${error?.message}")
